Added a fragment and navigation graph
