+++ app/src/main/res/navigation/kempton_park_aog_navigation_graph.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/kempton_park_aog_navigation_graph"
+    app:startDestination="@id/welcomeFragment">
+
+    <fragment
+        android:id="@+id/welcomeFragment"
+        android:name="com.example.kemptonparkaog.WelcomeFragment"
+        tools:layout="@layout/welcome_fragment">
+
+    </fragment>
+
+   <!-- <fragment
+        android:id="@+id/loginFragment"
+        android:name="com.example.kemptonparkaog.LoginFragment"
+        tools:layout="@layout/login_fragment">
+
+    </fragment>-->
+</navigation>
